Move all code out of the original StoredKey.kt file into a separate files in a package
